requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
fpdf2==2.4.6
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==9.0.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.2